Move PlayerArchetypeData interface to types.ts for improved organization and clarity
