+++ src/systems/types.ts
@@ -53,7 +53,35 @@ import type {
   DisplayObject as PixiDisplayObject,
   Texture,
 } from "pixi.js";
-import { PlayerArchetypeData } from "./vitalpoint/types";
+
+// Vital point effect
+// Player archetype data
+export interface PlayerArchetypeData {
+  readonly id: string;
+  readonly name: KoreanText;
+  readonly description: KoreanText;
+  readonly baseHealth: number;
+  readonly baseKi: number;
+  readonly baseStamina: number;
+  readonly coreStance: TrigramStance;
+  readonly theme: {
+    primary: number;
+    secondary: number;
+  };
+  readonly colors: {
+    primary: number;
+    secondary: number;
+  };
+  readonly stats: {
+    attackPower: number;
+    defense: number;
+    speed: number;
+    technique: number;
+  };
+  readonly favoredStances: readonly TrigramStance[];
+  readonly specialAbilities: readonly string[];
+  readonly philosophy: KoreanText;
+}
 
 export interface AnimationState {
   readonly currentAnimationName?: string;
